src/pyproject.toml
[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings_test"
pythonpath = ["."]

[project]
name = "testovoe"
version = "0.1.0"
description = "Тестовое задание"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django<6",
    "django-constance>=4.3.4",
    "django-environ==0.12.0",
    "django-imagekit==6.0.0",
    "django-ninja==1.5.3",
    "django-ninja-jwt==5.4.4",
    "django-post-office>=3.11.0",
    "django-tinymce>=5.0.0",
    "pillow==12.1.1",
    "psycopg2-binary==2.9.9",
    "redis>=7.1.1",
    "requests>=2.32.5",
    "openpyxl>=3.1.0",
]
